NET-17: move proofs to single file
